🐛 fix HTML displaying in error alerts
refs https://github.com/TryGhost/Ghost/issues/8203
- check for `<html>` element in error response details and if it's present fallback to the default error message
